+++ internal/auth/pam/pam.go
@@ -1,9 +1,5 @@
 //go:build pam
 
-// Copyright 2014 The Gogs Authors. All rights reserved.
-// Use of this source code is governed by a MIT-style
-// license that can be found in the LICENSE file.
-
 package pam
 
 import (