+++ mockgen.go
@@ -1,0 +1,4 @@
+package main
+
+//go:generate go install golang.org/x/tools/cmd/goimports@v0.17.0
+//go:generate go run github.com/unknwon/go-mockgen/cmd/go-mockgen@v0.0.0-20251002032800-a9a94b119e3b